move some lib files into subfolders for better ordering
