+++ src/lib/components/layout/Header.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
   // import Search from 'svelte-algolia'
   import { browser } from '$app/environment'
-  import { Nav, ThemeSwitcher } from '.'
+  import { Nav, ThemeSwitcher } from '../..'
   // import { SearchHit } from '.'
-  import type { NavLink } from './types'
+  import type { NavLink } from '../../types'
 
   interface Props {
     nav: NavLink[]
