+++ src/lib/components/ui/CircleSpinner.svelte
