+++ src/lib/index.ts
@@ -1,22 +1,29 @@
-export { default as BasePage } from './BasePage.svelte'
-export { default as ChapterList } from './ChapterList.svelte'
-export { default as ChapterMap } from './ChapterMap.svelte'
-export { default as CircleSpinner } from './CircleSpinner.svelte'
-export { default as Collapsible } from './Collapsible.svelte'
-export { default as Footer } from './Footer.svelte'
-export { default as FormField } from './FormField.svelte'
-export { default as Geocoder } from './Geocoder.svelte'
-export { default as Header } from './Header.svelte'
-export { default as Img } from './Img.svelte'
-export { default as Map } from './Map.svelte'
-export { default as Modal } from './Modal.svelte'
-export { default as Nav } from './Nav.svelte'
-export { default as PlaceSelect } from './PlaceSelect.svelte'
-export { default as PostPreview } from './PostPreview.svelte'
-export { default as RadioButtons } from './RadioButtons.svelte'
-export { default as SearchHit } from './SearchHit.svelte'
-export { default as Social } from './Social.svelte'
-export { default as TagList } from './TagList.svelte'
-export { default as ThemeSwitcher } from './ThemeSwitcher.svelte'
-export { default as Toggle } from './Toggle.svelte'
-export { default as ToolTip } from './ToolTip.svelte'
+// Layout components
+export { default as BasePage } from './components/layout/BasePage.svelte'
+export { default as Footer } from './components/layout/Footer.svelte'
+export { default as Header } from './components/layout/Header.svelte'
+export { default as Nav } from './components/layout/Nav.svelte'
+
+// Form components
+export { default as Collapsible } from './components/forms/Collapsible.svelte'
+export { default as FormField } from './components/forms/FormField.svelte'
+export { default as PlaceSelect } from './components/forms/PlaceSelect.svelte'
+export { default as RadioButtons } from './components/forms/RadioButtons.svelte'
+export { default as Toggle } from './components/forms/Toggle.svelte'
+
+// Map components
+export { default as ChapterMap } from './components/maps/ChapterMap.svelte'
+export { default as Geocoder } from './components/maps/Geocoder.svelte'
+export { default as Map } from './components/maps/Map.svelte'
+
+// UI components
+export { default as ChapterList } from './components/ui/ChapterList.svelte'
+export { default as CircleSpinner } from './components/ui/CircleSpinner.svelte'
+export { default as Img } from './components/ui/Img.svelte'
+export { default as Modal } from './components/ui/Modal.svelte'
+export { default as PostPreview } from './components/ui/PostPreview.svelte'
+export { default as SearchHit } from './components/ui/SearchHit.svelte'
+export { default as Social } from './components/ui/Social.svelte'
+export { default as TagList } from './components/ui/TagList.svelte'
+export { default as ThemeSwitcher } from './components/ui/ThemeSwitcher.svelte'
+export { default as ToolTip } from './components/ui/ToolTip.svelte'
